Object.Is() is now only use in other class for defining Object name. And removed objPtrDic and use objPtrTuple instead. For less complexity of code. Object name will define by default using same data in Map.registerObject("Name").
in update function in gameEngine class, Added Event checker that will check for position of every Object in the game, and sort queue for Object to draw on screen. So now game is now capable for Object depth. If Object is a bit on above of another Object. It will be detect that this Object is behide the another Object and it will be kept for drawing later. Waiting for frontline Object to be drawn first.